added chat component
